+++ src/items/decay/decay.h
@@ -44,7 +44,7 @@ class Decay
 		void checkDecay();
 		void internalDecayItem(Item* item);
 
-		uint64_t eventId {0};
+		uint32_t eventId {0};
 		std::map<int64_t, std::vector<Item*>> decayMap;
 };
